+++ python_ta/checkers/one_iteration_checker.py
@@ -19,6 +19,17 @@ class OneIterationChecker(BaseChecker):
             '(e.g., by returning or using the "break" keyword).',
         )
     }
+    options = (
+        (
+            "z3",
+            {
+                "default": False,
+                "type": "yn",
+                "metavar": "<y or n>",
+                "help": "Use Z3 to restrict control flow checks to paths that are logically feasible.",
+            },
+        ),
+    )
 
     @only_required_for_messages("one-iteration")
     def visit_for(self, node: nodes.For) -> None:
@@ -52,6 +63,19 @@ class OneIterationChecker(BaseChecker):
 
         if preds == []:
             return False
+
+        if self.linter.config.z3:
+            # check whether the loop statement is feasible
+            if not start.cfg_block.is_feasible:
+                return False
+            # check whether the loop body is feasible (the loop has at least one iteration)
+            if not any(
+                succ.is_feasible and node.parent_of(succ.target.statements[0])
+                for succ in start.cfg_block.successors
+            ):
+                return False
+            # filter out edges from unfeasible blocks
+            preds = [pred for pred in preds if pred.source.is_feasible]
 
         for pred in preds:
             stmt = pred.source.statements[0]